refactor code

+++ commentthread/comment/comment_service.go
@@ -4,15 +4,16 @@ import (
 	"context"
 	"database/sql"
 	"database/sql/driver"
+	"errors"
 	"fmt"
 	"time"
 )
 
 type CommentService interface {
-	GetComments(ctx context.Context, commentThreadId string, userId *string) ([]Comment, error)
-	Create(ctx context.Context, comment Comment) (int64, error)
-	Update(ctx context.Context, comment Comment) (int64, error)
-	Remove(ctx context.Context, commentId string, commentThreadId string) (int64, error)
+	GetComments(ctx context.Context, commentThreadId string, userId *string) ([]Response, error)
+	Create(ctx context.Context, id, commentId, commentThreadId, author string, comment Request) (int64, error)
+	Update(ctx context.Context, commentId, author string, comment Request) (int64, error)
+	Remove(ctx context.Context, commentId string, commentThreadId string, author string) (int64, error)
 }
 
 func NewCommentService(db *sql.DB, replyTable string, commentIdCol string, authorCol string, idCol string, updatedAtCol string, commentCol string, userIdCol string, timeCol string, historiesCol string, commentThreadIdCol string, reactionCol string, commentReactionTable string, commentIdReactionCol string, userTable string, userIdUserCol string, usernameUserCol string, avatarUserCol string, commentInfoTable string, userfulCountInfoCol string, commentIdInfoCol string, commentThreadInfoTable string, commentIdCommentThreadInfoCol string, replyCountCommentThreadInfoCol string, usefulCountCommentThreadInfoCol string, queryInfo func(ids []string) ([]Info, error), toArray func(interface{}) interface {
@@ -28,8 +29,8 @@ func NewCommentService(db *sql.DB, replyTable string, commentIdCol string, autho
 		updatedAtCol:                    updatedAtCol,
 		historiesCol:                    historiesCol,
 		commentCol:                      commentCol,
-		reactionCol:                     reactionCol,
 		userIdCol:                       userIdCol,
+		reactionCol:                     reactionCol,
 		timeCol:                         timeCol,
 		commentThreadIdCol:              commentThreadIdCol,
 		commentReactionTable:            commentReactionTable,
@@ -84,13 +85,16 @@ type commentService struct {
 	}
 }
 
-func (s *commentService) Update(ctx context.Context, comment Comment) (int64, error) {
-	qr := fmt.Sprintf("select %s, %s,%s from %s where %s = $1", s.commentIdCol, s.commentCol, s.historiesCol, s.ReplyTable, s.commentIdCol)
-	rows := s.db.QueryRow(qr, comment.CommentId)
+func (s *commentService) Update(ctx context.Context, commentId, author string, req Request) (int64, error) {
+	qr := fmt.Sprintf("select %s, %s,%s,%s from %s where %s = $1", s.commentIdCol, s.commentCol, s.historiesCol, s.authorCol, s.ReplyTable, s.commentIdCol)
+	rows := s.db.QueryRow(qr, commentId)
 	var exist = Comment{}
-	err := rows.Scan(&exist.CommentId, &exist.Comment, s.toArray(&exist.Histories))
+	err := rows.Scan(&exist.CommentId, &exist.Comment, s.toArray(&exist.Histories), exist.Author)
 	if err != nil {
 		return -1, err
+	}
+	if exist.Author == "" || exist.Author != author {
+		return -2, errors.New("no permission on comment")
 	}
 	updatedTime := time.Now()
 	exist.Histories = append(exist.Histories, History{
@@ -98,7 +102,7 @@ func (s *commentService) Update(ctx context.Context, comment Comment) (int64, er
 		Time:    updatedTime,
 	})
 	qr1 := fmt.Sprintf("update %s set %s = $1, %s = $2, %s = $3", s.ReplyTable, s.commentCol, s.historiesCol, s.updatedAtCol)
-	rows2, er2 := s.db.ExecContext(ctx, qr1, comment.Comment, s.toArray(exist.Histories), updatedTime)
+	rows2, er2 := s.db.ExecContext(ctx, qr1, req.Comment, s.toArray(exist.Histories), updatedTime)
 	if er2 != nil {
 		return -1, er2
 	}
@@ -106,7 +110,15 @@ func (s *commentService) Update(ctx context.Context, comment Comment) (int64, er
 }
 
 // Delete implements CommentThreadReplyService
-func (s *commentService) Remove(ctx context.Context, commentId string, commentThreadId string) (int64, error) {
+func (s *commentService) Remove(ctx context.Context, commentId string, commentThreadId string, author string) (int64, error) {
+	count := 0
+	err := s.db.QueryRow(fmt.Sprintf("SELECT COUNT(*) FROM %s WHERE %s=$1 and %s=$2", s.ReplyTable, s.commentIdCol, s.authorCol), commentId, author).Scan(&count)
+	if err != nil {
+		return 0, err
+	}
+	if count == 0 {
+		return 0, errors.New("user does not have permission to delete the comment")
+	}
 	rowsAffected := int64(0)
 	qr1 := fmt.Sprintf("delete from %s where %s = $1", s.ReplyTable, s.commentIdCol)
 	tx, err := s.db.BeginTx(ctx, nil)
@@ -160,7 +172,8 @@ func (s *commentService) Remove(ctx context.Context, commentId string, commentTh
 }
 
 // Create implements CommentThreadReplyService
-func (s *commentService) Create(ctx context.Context, comment Comment) (int64, error) {
+func (s *commentService) Create(ctx context.Context, id, commentId, commentThreadId, author string, req Request) (int64, error) {
+	comment := Comment{Id: id, CommentId: commentId, CommentThreadId: commentThreadId, Author: author, Comment: req.Comment, Time: time.Now()}
 	rowsAffected := int64(0)
 	tx, err := s.db.BeginTx(ctx, nil)
 	if err != nil {
@@ -168,9 +181,9 @@ func (s *commentService) Create(ctx context.Context, comment Comment) (int64, er
 	}
 	defer tx.Rollback()
 
-	qr := fmt.Sprintf("insert into %s(%s,%s,%s,%s,%s,%s,%s,%s) values($1,$2,$3,$4,$5,$6,$7,$8)",
-		s.ReplyTable, s.commentIdCol, s.idCol, s.authorCol, s.userIdCol, s.commentCol, s.timeCol, s.historiesCol, s.commentThreadIdCol)
-	rows1, err := tx.ExecContext(ctx, qr, comment.CommentId, comment.Id, comment.Author, comment.UserId, comment.Comment, time.Now(), s.toArray([]interface{}{}), comment.CommentThreadId)
+	qr := fmt.Sprintf("insert into %s(%s,%s,%s,%s,%s,%s,%s) values($1,$2,$3,$4,$5,$6,$7)",
+		s.ReplyTable, s.commentIdCol, s.idCol, s.authorCol, s.commentCol, s.timeCol, s.historiesCol, s.commentThreadIdCol)
+	rows1, err := tx.ExecContext(ctx, qr, comment.CommentId, comment.Id, comment.Author, comment.Comment, time.Now(), s.toArray([]interface{}{}), comment.CommentThreadId)
 	if err != nil {
 		return -1, err
 	}
@@ -206,14 +219,15 @@ func (s *commentService) Create(ctx context.Context, comment Comment) (int64, er
 	return rowsAffected, nil
 }
 
-func (s *commentService) GetComments(ctx context.Context, commentThreadId string, userId *string) ([]Comment, error) {
+func (s *commentService) GetComments(ctx context.Context, commentThreadId string, userId *string) ([]Response, error) {
 	qr := ""
 	qr2 := ""
+	rs := make([]Response, 0)
 	arr := []interface{}{}
 	if userId != nil && len(*userId) > 0 {
 		arr = append(arr, userId)
 		qr = fmt.Sprintf(`, case when d.%s = 1 then true else false end as disable`, s.reactionCol)
-		qr2 = fmt.Sprintf(`left join %s d on a.%s = d.%s and a.%s = $1`,
+		qr2 = fmt.Sprintf(`left join %s d on a.%s = d.%s and d.%s = $1`,
 			s.commentReactionTable, s.commentIdCol, s.commentIdReactionCol, s.userIdCol)
 	}
 	param := "1"
@@ -240,13 +254,11 @@ func (s *commentService) GetComments(ctx context.Context, commentThreadId string
 			&comment.CommentThreadId,
 			&comment.Id,
 			&comment.Author,
-			&comment.UserId,
 			&comment.Comment,
 			&comment.Time,
 			&comment.UpdatedAt,
 			s.toArray(&comment.Histories),
 			&comment.UsefulCount,
-			&comment.Disable,
 		)
 		if err != nil {
 			return nil, err
@@ -255,11 +267,11 @@ func (s *commentService) GetComments(ctx context.Context, commentThreadId string
 		comments = append(comments, comment)
 	}
 	if len(comments) == 0 {
-		return comments, nil
+		return rs, nil
 	}
 	ids := make([]string, 0)
 	for _, r := range comments {
-		ids = append(ids, r.UserId)
+		ids = append(ids, r.Author)
 	}
 	infos, err := s.queryInfo(ids)
 	if err != nil {
@@ -267,15 +279,16 @@ func (s *commentService) GetComments(ctx context.Context, commentThreadId string
 	}
 	for k, _ := range comments {
 		c := comments[k]
-		i := BinarySearch(infos, c.UserId)
+		r := toResponse(c)
+		i := BinarySearch(infos, c.Author)
 		if i >= 0 {
-			if comments[k].UserId == infos[i].Id {
-				comments[k].AuthorURL = &infos[i].Url
-				comments[k].AuthorName = &infos[i].Name
+			if comments[k].Author == infos[i].Id {
+				r.AuthorURL = &infos[i].Url
+				r.AuthorName = &infos[i].Name
 			}
-
 		}
+		rs = append(rs, r)
 	}
 
-	return comments, nil
+	return rs, nil
 }
